requirements.txt
# Django Core
Django==5.1.5
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Caching & Queue
django-redis==5.4.0
redis==5.0.1
celery[redis]==5.4.0
flower==2.0.1

# Authentication & Permissions
django-allauth==65.3.0
django-guardian==2.4.0

# Forms & UI
django-crispy-forms==2.3
crispy-bootstrap5==0.7

# Admin Enhancement
django-baton==4.1.0

# Search
django-haystack==3.3.0

# Filtering
django-filter==24.3

# Storage & Files
django-storages==1.14.4
Pillow==10.4.0

# Import/Export
django-import-export==4.3.3

# Utilities
django-extensions==3.2.3
django-safedelete==1.4.0

# Money handling
django-money==3.5.3
py-moneyed==3.0

# Template Engine
Jinja2==3.1.4

# Development Tools
django-debug-toolbar==4.4.6

# API (for future use)
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Static Files
whitenoise==6.7.0

# Production Server
gunicorn==23.0.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1

# Code Quality
black==24.10.0
flake8==7.1.1
isort==5.13.2

